requirements.txt
langchain
google-generativeai==0.3.1
chromadb
pypdfium2
langchain-google-genai
streamlit


